+++ docs/changelog.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [v0.3.0]
 
 ### Added
 
@@ -12,6 +12,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Get strings by substrings
 - Get links by contents, content substrings
 - Add opportunity to create sc-elements by SCs-text
+
+### Changed
+
+- Revert of "Add params for search" revert
+- Revert of "Add opportunity to pass template and params by addr and idtf" revert
 
 ### Fixed